requirements.txt
numpy==1.18.3
Flask==1.1.1
gunicorn==20.0.4
google-cloud-datastore==1.12.0